requirements.test.txt
factory_boy
pytest
pytest-django
pytest-profiling
pytest-random-order
pytest-xdist
